key geeration complete

+++ src/prime.rs
@@ -2,7 +2,7 @@
 //! This module generates and fills a vector with prime numbers for rsa
 
 ///Generate an vector of prime numbers and return the vector
-pub fn generate_primes(n: i32) -> Result<Vec<i32>, String> {
+pub fn generate_primes(n: i64) -> Result<Vec<i64>, String> {
     let mut primes = Vec::new();
     if n <= 0 {
         return Err(String::from("The end of range must be a postive integer"));
@@ -16,11 +16,11 @@ pub fn generate_primes(n: i32) -> Result<Vec<i32>, String> {
 }
 
 ///Check if number is prime
-fn is_prime(n: i32) -> bool {
+fn is_prime(n: i64) -> bool {
     if n <= 1 {
         return false;
     }
-    let end = (n as f64).sqrt() as i32 + 1;
+    let end = (n as f64).sqrt() as i64 + 1;
     for i in 2..end {
         if n % i == 0 {
             return false;
